Laravel E-commerce Part-8

+++ database/migrations/2019_08_17_062138_create_categories_table.php
@@ -17,7 +17,7 @@ class CreateCategoriesTable extends Migration
             $table->string('name');
             $table->string('slug')->unique();
             $table->text('description')->nullable();
-            $table->unsignedInteger('parent_id')->default(1)->nullable();
+            $table->unsignedBigInteger('parent_id')->default(1)->nullable();
             $table->boolean('featured')->default(0);
             $table->boolean('menu')->default(1);
             $table->string('image')->nullable();
